requirements.txt
Flask
keras
flask_restful
scikit-learn
numpy
pandas
seaborn
matplotlib.pyplot
tensorflow
joblib
pickle
json